# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_inbound_http_server.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_eng-yousef-khaled_expenses_api_internal_core_auth.User"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_eng-yousef-khaled_expenses_api_internal_core_auth.Password": {
            "type": "object",
            "properties": {
                "hashed": {
                    "type": "boolean"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "github_com_eng-yousef-khaled_expenses_api_internal_core_auth.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "isVerification": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "$ref": "#/definitions/github_com_eng-yousef-khaled_expenses_api_internal_core_auth.Password"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "internal_inbound_http_server.CreateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        }
    }
}